First commit
This first version implements all the CPU instructions plus a very basic
version of the memory system.

This first version passes 10 out of the 11 Blargg's CPU tests. There's a
test that fails because interrupts are not implemented yet.

Blargg's test ROMs are hosted here:
http://gbdev.gg8.se/files/roms/blargg-gb-tests/

There are some things that could be simplified, but that's OK. The
important thing for now is that most of the CPU tests are passing and
so the current code is a good base to keep working on the rest of
subsystems.